third_party/blink/web_tests/external/wpt/wasm/jsapi/instance/constructor-caching.any-expected.txt

This is a testharness.js-based test.
FAIL constructor-caching assert_equals: expected object "[object WebAssembly.Global]" but got object "[object WebAssembly.Global]"
Harness: the test ran to completion.
